Move get transactions history test to other file
